requirements.txt
Django==1.10.1
argparse==1.2.1
django-compressor==2.1
wsgiref==0.1.2
djangorestframework==3.4.6
drf-nested-routers==0.11.1